move workout-related scripts into new workouts folder
