+++ frontend/www/exercise.html
@@ -95,7 +95,7 @@
         <div class="col-lg-6"></div>
       </form>
     </div>
-    <script type="module" src="scripts/exercise.js"></script>
+    <script type="module" src="scripts/workouts/exercise.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
