+++ frontend/www/exercises.html
@@ -79,7 +79,7 @@
       </a>
     </template>
 
-    <script type="module" src="scripts/exercises.js"></script>
+    <script type="module" src="scripts/workouts/exercises.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
