+++ frontend/www/gallery.html
@@ -41,7 +41,7 @@
       </div>
     </div>
 
-    <script type="module" src="scripts/gallery/gallery.js"></script>
+    <script type="module" src="scripts/workouts/gallery/gallery.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
