+++ frontend/www/scripts/workouts/exercise.js
@@ -1,6 +1,6 @@
-import { setReadOnly, createAlert } from "./utils/dom.js";
-import { sendRequest } from "./utils/api.js";
-import { HOST } from "./utils/host.js";
+import { setReadOnly, createAlert } from "../utils/dom.js";
+import { sendRequest } from "../utils/api.js";
+import { HOST } from "../utils/host.js";
 
 let cancelButton;
 let okButton;
