+++ frontend/www/scripts/workouts/workout.js
@@ -1,6 +1,6 @@
-import { sendRequest, getCurrentUser } from "./utils/api.js";
-import { HOST } from "./utils/host.js";
-import { createAlert, setReadOnly } from "./utils/dom.js";
+import { sendRequest, getCurrentUser } from "../utils/api.js";
+import { HOST } from "../utils/host.js";
+import { createAlert, setReadOnly } from "../utils/dom.js";
 
 let cancelWorkoutButton;
 let okWorkoutButton;
