+++ frontend/www/workout.html
@@ -168,7 +168,7 @@
       </div>
     </template>
 
-    <script type="module" src="scripts/workout.js"></script>
+    <script type="module" src="scripts/workouts/workout.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-ygbV9kiqUc6oa4msXn9868pTtWMgiQaeYH7/t7LECLbyPA2x65Kgf80OJFdroafW"
